+++ include/Utils.hpp
@@ -11,14 +11,14 @@ namespace sat::utils {
 class RandomGen {
  public:
   static std::random_device rd;
-  static unsigned int initialSeed;
+  static unsigned long initialSeed;
   static std::mt19937 randomGenerator;
   static std::uniform_int_distribution<> randomBoolUD;
   static std::uniform_real_distribution<> randomReal01UD;
 
   inline static void setSeed(int seed) { randomGenerator.seed(seed); }
   inline static bool getRandomBool() { return randomBoolUD(randomGenerator); }
-  inline static double getRandomReal01() {
+  inline static float getRandomReal01() {
     return randomReal01UD(randomGenerator);
   }
 };
